requirements.txt
# Remote API client
requests>=2.31.0

# LLM
openai>=1.0.0
pyyaml>=6.0

# Workflow orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# Optional: litellm support
# litellm>=1.0.0

# Utilities
python-dotenv>=1.0.0
